+++ src/arcIO.NET/Assay.fs
@@ -13,6 +13,10 @@ module Assay =
     let subFolderPaths = 
         ["dataset";"protocol"]
 
+    let nameToFileName (n:string) =
+        if n.Contains("/") || n.Contains("\\") then n else
+        Path.Combine(n,assayFileName)
+
     module AssayFolder =
         
         /// Checks if an assay folder exists in the ARC.
@@ -21,18 +25,33 @@ module Assay =
             |> System.IO.Directory.Exists
 
     let readFromFolder (folderPath : string) =
-        let ap = Path.Combine (folderPath,assayFileName)
+        let ap = Path.Combine(folderPath,assayFileName).Replace(@"\","/")
         let c,a = AssayFile.Assay.fromFile ap
         c,a
 
     let readByFileName (arc : string) (assayFileName : string) =
-        let ap = Path.Combine ([|arc;rootFolderName;assayFileName|])
+        let ap = Path.Combine([|arc;rootFolderName;assayFileName|]).Replace(@"\","/")
         let c,a = AssayFile.Assay.fromFile ap
         c,a
 
     let readByName (arc : string) (assayName : string) =
-        Path.Combine ([|arc;rootFolderName;assayName|])
+        Path.Combine([|arc;rootFolderName;assayName|]).Replace(@"\","/")
         |> readFromFolder
+
+    let tryReadFromFolder (folderPath : string) =
+        try 
+            readFromFolder folderPath |> Some
+        with | _ -> None
+
+    let tryReadByFileName (arc : string) (assayFileName : string) =
+        try 
+            readByFileName arc assayFileName |> Some
+        with | _ -> None
+
+    let tryReadByName (arc : string) (assayName : string) =
+        try 
+            readByName arc assayName |> Some
+        with | _ -> None
 
     let writeToFolder (folderPath : string) (contacts : Person list) (assay : Assay) =
         let ap = Path.Combine (folderPath,assayFileName)
@@ -71,9 +90,16 @@ module Assay =
 
 
     let initFromName (arc : string) (assayName : string) =
-        
-        let assayFileName = Path.Combine(assayName,assayFileName)
 
-        let assay = Assay.create(FileName = assayFileName)
+        let assay = Assay.create(FileName = nameToFileName assayName)
 
         init arc assay
+
+[<AutoOpen>]
+module AssayExtensions =
+    type Assay with
+        static member create (?Id,?Name,?MeasurementType,?TechnologyType,?TechnologyPlatform,?DataFiles,?Materials,?CharacteristicCategories,?UnitCategories,?ProcessSequence,?Comments) : Assay =
+            let Filename = 
+                Name
+                |> Option.map Assay.nameToFileName
+            Assay.make Id Filename MeasurementType TechnologyType TechnologyPlatform DataFiles Materials CharacteristicCategories UnitCategories ProcessSequence Comments