+++ src/arcIO.NET/Investigation.fs
@@ -8,9 +8,40 @@ module Investigation =
 
     let investigationFileName = "isa.investigation.xlsx"
 
-    let fromArcFolder (arc : string) =
+    /// Creates an investigation file in the ARC.
+    let write (arcDir : string) (investigation : ISADotNet.Investigation) =
+           
+        let log = Logging.createLogger "InvestigationWriteLog"
+        
+        log.Info("Start Investigation Write")
+
+        if System.IO.File.Exists(Path.Combine(arcDir,investigationFileName)) then
+            log.Error("Investigation file does already exist.")
+
+        else 
+            let investigationFilePath = Path.Combine(arcDir,investigationFileName)    
+            Investigation.toFile investigationFilePath investigation
+
+    /// Reads an investigation from the ARC.
+    let read (arcDir : string) =
+           
+        let log = Logging.createLogger "InvestigationReadLog"
+        
+        log.Info("Start Investigation Read")
+
+        if System.IO.File.Exists(Path.Combine(arcDir,investigationFileName)) then
+            let investigationFilePath = Path.Combine(arcDir,investigationFileName)    
+            Investigation.fromFile investigationFilePath
+
+        else 
+            log.Error("Investigation file does not exist.")
+            raise (System.SystemException())
+
+    let fromArcFolder (arcDir : string) =
+        let log = Logging.createLogger "InvestigationFromArcFolderLog"
+
         // read investigation from investigation file
-        let ip = Path.Combine(arc,investigationFileName)
+        let ip = Path.Combine(arcDir,investigationFileName).Replace(@"\","/")
         let i = Investigation.fromFile ip
 
         // get study list from study files and assay files
@@ -18,22 +49,32 @@ module Investigation =
             i.Studies
             |> Option.map (List.map (fun study -> 
                 // read study from file
-                let studyFromFile = Study.readByIdentifier arc study.Identifier.Value
-                // update study assays and contacts with information from assay files
-                match study.Assays with
-                | Some assays ->
-                    let scontacts,sassays = 
-                        assays
-                        |> List.fold (fun (cl,al) assay ->
-                            let contactsFromFile,assayFromFile = Assay.readByFileName arc assay.FileName.Value
-                            cl @ contactsFromFile, al @ [assayFromFile]
-                        ) (studyFromFile.Contacts |> Option.defaultValue [],[])
-                    {studyFromFile with                        
-                        Contacts = Some scontacts
-                        Assays = Some sassays 
-                    }
-                | None -> 
-                    studyFromFile
+                match study.Identifier with
+                | Some id ->
+                    let studyFromFile = Study.readByIdentifier arcDir id
+                    // update study assays and contacts with information from assay files
+                    match study.Assays with
+                    | Some assays ->
+                        let scontacts,sassays = 
+                            assays
+                            |> List.fold (fun (cl,al) assay ->
+                                match assay.FileName with
+                                | Some fn ->
+                                    let contactsFromFile,assayFromFile = Assay.readByFileName arcDir assay.FileName.Value
+                                    cl @ contactsFromFile, al @ [assayFromFile]
+                                | None ->
+                                    log.Warn("Study \'" + id + "\' contains Assay without filename.")
+                                    cl, al @ [assay]
+                            ) (studyFromFile.Contacts |> Option.defaultValue [],[])
+                        {studyFromFile with                        
+                            Contacts = Some (scontacts |> List.distinct)
+                            Assays = Some sassays 
+                        }
+                    | None -> 
+                        studyFromFile
+                | None ->
+                    log.Warn("Investigation file contains study without identifier.")
+                    study
             ))
         
         // construct complete process list from studies and assays, then update by itself
@@ -66,6 +107,39 @@ module Investigation =
         // fill investigation with information from study files and assay files
         {i with Studies = istudies'}
 
-    let write (arc : string) (investigation : Investigation) =
-        let p = Path.Combine(arc,investigationFileName)
-        ISADotNet.XLSX.Investigation.toFile p investigation
+    let registerAssay arcDir studyName (assayName) =
+
+        let log = Logging.createLogger "RegisterAssayLog"
+
+        let _, assay = Assay.readByName arcDir assayName
+
+        let investigation = read arcDir
+
+        match investigation.Studies with
+        | Some studies -> 
+            match API.Study.tryGetByIdentifier studyName studies with
+            | Some study -> 
+                match study.Assays with
+                | Some assays -> 
+                    match Assay.tryReadByName arcDir assayName with
+                    | Some _ ->
+                        log.Error($"Assay with the identifier {assayName} already exists in the investigation file.")
+                        assays
+                    | None ->
+                        API.Assay.add assays assay
+                | None ->
+                    [assay]
+                |> API.Study.setAssays study
+                |> fun s -> API.Study.updateByIdentifier API.Update.UpdateAll s studies
+            | None ->
+                log.Info($"Study with the identifier {studyName} does not exist yet, creating it now.")
+                let study = Study.create(Identifier = studyName, Assays = [assay])
+                Study.init arcDir study
+                API.Study.add studies study
+        | None ->
+            log.Info($"Study with the identifier {studyName} does not exist yet, creating it now.")
+            let study = Study.create(Identifier = studyName, Assays = [assay])
+            Study.init arcDir study
+            [study]
+        |> API.Investigation.setStudies investigation
+        |> write arcDir