+++ RELEASE_NOTES.md
@@ -1,3 +1,15 @@
+### 0.0.6+da8c364 (Released 2023-1-26)
+* Additions:
+    * [[#6ed4bc8](https://github.com/nfdi4plants/arcIO.NET/commit/6ed4bc840dc09e2f8dd348a79241eb64c7de8d5c)] migrate githelper and logging from arcCommander
+    * [[#929bab1](https://github.com/nfdi4plants/arcIO.NET/commit/929bab19595e511d96367a09945fab943638c13e)] Improve linux file reads for Study and Assay: autocorrect backslashes to frontslashes
+    * [[#33e6b26](https://github.com/nfdi4plants/arcIO.NET/commit/33e6b267d3ec7d18274a47e2e8420b1a258af67a)] Improved investigation fromArcFolder and added autocorrect for reading file paths.
+    * [[#23005bc](https://github.com/nfdi4plants/arcIO.NET/commit/23005bce740a7f0a4cd1f2d36c372d556282069f)] Assay create by name instead of create by filename
+    * [[#bfa0335](https://github.com/nfdi4plants/arcIO.NET/commit/bfa0335cd7607d8fd471080f37e4eaffde7eb2bc)] Added arc init functionality (subfolders, investigation file, git)
+    * [[#cf082e2](https://github.com/nfdi4plants/arcIO.NET/commit/cf082e2842f21b01503b453e3d17dde308bb8b16)] Added assay register to Investigation module
+    * [[#da8c364](https://github.com/nfdi4plants/arcIO.NET/commit/da8c36448bec084fec3fc5fe7a08d6aa55a92521)] update isadotnet reference
+* Bugfixes:
+    * [[#b4748be](https://github.com/nfdi4plants/arcIO.NET/commit/b4748bef60b91b2dc148fe4d3018733f5e810790)] build fixes for linux
+
 ### 0.0.5+1d873f4 (Released 2023-1-8)
 * Additions:
     * [[#b7c205e](https://github.com/nfdi4plants/arcIO.NET/commit/b7c205e6dd6a219d1c3f9b4dbe8ca7d1883785f2)] update ISADotNet reference